Move the implementation of the new session queue into the local package
